Implement controller, service, repository and db delegate for AB#15406

+++ Apps/Database/src/Delegates/DbBlockedAccessDelegate.cs
@@ -18,6 +18,7 @@ namespace HealthGateway.Database.Delegates
     using System.Collections.Generic;
     using System.Linq;
     using System.Threading.Tasks;
+    using HealthGateway.Common.Data.Constants;
     using HealthGateway.Database.Context;
     using HealthGateway.Database.Models;
     using Microsoft.EntityFrameworkCore;
@@ -43,7 +44,15 @@ namespace HealthGateway.Database.Delegates
         /// <inheritdoc/>
         public async Task DeleteBlockedAccessAsync(BlockedAccess blockedAccess, AgentAudit agentAudit)
         {
-            this.dbContext.BlockedAccess.Remove(blockedAccess);
+            this.logger.LogDebug("Blocked access version: {Version} for hdid: {Hdid}", blockedAccess.Version, blockedAccess.Hdid);
+
+            // Only attempt to remove entity if version is not 0
+            if (blockedAccess.Version != 0)
+            {
+                this.dbContext.BlockedAccess.Remove(blockedAccess);
+                this.logger.LogDebug("Blocked access removed for Hdid: {Hdid}", blockedAccess.Hdid);
+            }
+
             this.dbContext.AgentAudit.Add(agentAudit);
 
             await this.dbContext.SaveChangesAsync().ConfigureAwait(true);
@@ -52,8 +61,6 @@ namespace HealthGateway.Database.Delegates
         /// <inheritdoc/>
         public async Task<BlockedAccess?> GetBlockedAccessAsync(string hdid)
         {
-            this.logger.LogTrace("Getting blocked access for hdid: {Hdid}", hdid);
-
             IQueryable<BlockedAccess> query = this.dbContext.BlockedAccess
                 .Where(d => d.Hdid == hdid);
 
@@ -61,14 +68,12 @@ namespace HealthGateway.Database.Delegates
         }
 
         /// <inheritdoc/>
-        public async Task<Dictionary<string, string>> GetDataSourcesAsync(string hdid)
+        public async Task<IEnumerable<DataSource>> GetDataSourcesAsync(string hdid)
         {
-            this.logger.LogTrace("Getting data sources from blocked access for hdid : {Hdid}", hdid);
-
             IQueryable<BlockedAccess> query = this.dbContext.BlockedAccess.Where(d => d.Hdid == hdid);
             BlockedAccess? blockedAccess = await query.SingleOrDefaultAsync().ConfigureAwait(true);
 
-            return blockedAccess?.DataSources ?? new Dictionary<string, string>();
+            return blockedAccess?.DataSources ?? new HashSet<DataSource>();
         }
 
         /// <inheritdoc/>
